+++ internal/authz/kessel/kessel.go
@@ -3,6 +3,7 @@ package kessel
 import (
 	"context"
 	"fmt"
+
 	"google.golang.org/protobuf/proto"
 
 	"github.com/spf13/viper"
