+++ internal/service/resources/kesselinventoryservice_test.go
@@ -1,4 +1,4 @@
-package kessel
+package resources
 
 import (
 	"github.com/stretchr/testify/assert"
